Moved progressive_cuts, Nminus1, etc... to named_func_utilities.cpp and fixed dependencies on all files (I believe).
